:truck: Extract utils into util-dir
