+++ @navikt/core/react/src/util/link-anchor/index.tsx
@@ -1,0 +1,7 @@
+"use client";
+export { LinkAnchor, LinkAnchorOverlay, LinkAnchorArrow } from "./LinkAnchor";
+export type {
+  LinkAnchorProps,
+  LinkAnchorOverlayProps,
+  LinkAnchorArrowProps,
+} from "./LinkAnchor";
